refactor #2 - extract out aws utils
